Extract common behaviour into a util
